requirements.txt
fastapi
uvicorn
loguru
google-cloud-bigquery
pandas
polars
python-dotenv
python-multipart
pyarrow